+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: hsilverb <hsilverb@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 15:49:06 by hsilverb          #+#    #+#             */
-/*   Updated: 2022/11/28 14:40:40 by hsilverb         ###   ########lyon.fr   */
+/*   Updated: 2022/11/29 17:09:50 by hsilverb         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
